backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.0
pypdf==3.17.0
python-multipart==0.0.6
pydantic==2.5.0
langchain==0.0.340
langchain-openai==0.0.2